# [FIX] product: speed up _name_search when variants are disabled