## Before this commit:
`_name_search` would scan the entire `product.product` table to find templates without any corresponding variants, which is expensive for databases with numerous products.

## After this commit:
`_name_search` skips templates without variants when variants are disabled, as each template should have exactly one variant.

### Benchmark for ~600k products
| Before | After |
|--------|--------|
| 1.837 s | 0.369s | 

opw-3489413