@robodoo 2023-10-06
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/137878).